requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
Django==4.1.7
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg2==1.19.4
feedparser==6.0.10
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
psycopg==3.1.12
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
schedule==1.2.1
sgmllib3k==1.0.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
v==1
